move top navigate into individual file because it will be fatter.
