+++ osu.Game.Rulesets.Karaoke/Edit/ImportLyric/AssignLanguage/AssignLanguageStepScreen.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Sprites;
@@ -75,41 +74,6 @@ namespace osu.Game.Rulesets.Karaoke.Edit.ImportLyric.AssignLanguage
                 lyricLanguageChangeHandler.AutoGenerate();
                 Navigation.State = NavigationState.Done;
             }));
-        }
-
-        public class AssignLanguageNavigation : TopNavigation<AssignLanguageStepScreen>
-        {
-            private const string auto_assign_language = "AUTO_ASSIGN_LANGUAGE";
-
-            public AssignLanguageNavigation(AssignLanguageStepScreen screen)
-                : base(screen)
-            {
-            }
-
-            protected override NavigationTextContainer CreateTextContainer()
-                => new AssignLanguageTextFlowContainer(Screen);
-
-            protected override void UpdateState(NavigationState value)
-            {
-                base.UpdateState(value);
-
-                NavigationText = value switch
-                {
-                    NavigationState.Initial => $"Try to select left side to mark lyric's language, or click [{auto_assign_language}] to let system auto detect lyric language.",
-                    NavigationState.Working => $"Almost there, you can still click [{auto_assign_language}] to re-detect each lyric's language.",
-                    NavigationState.Done => "Cool! Seems all lyric has it's own language. Go to next step to generate ruby.",
-                    NavigationState.Error => "Oops, seems cause some error in here.",
-                    _ => throw new ArgumentOutOfRangeException(nameof(value))
-                };
-            }
-
-            private class AssignLanguageTextFlowContainer : NavigationTextContainer
-            {
-                public AssignLanguageTextFlowContainer(AssignLanguageStepScreen screen)
-                {
-                    AddLinkFactory(auto_assign_language, "language detector", screen.AskForAutoAssignLanguage);
-                }
-            }
         }
     }
 }
